requirements.txt
playwright==1.44.0
apscheduler==3.10.4
google-auth==2.29.0
google-api-python-client==2.128.0
python-dotenv==1.0.1
